% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crop_P_K_demand.R
\name{calc_crop_K_demand}
\alias{calc_crop_K_demand}
\title{Crop K2O demand}
\usage{
calc_crop_K_demand(
  expected_yield_tons_ha,
  crop,
  uptake_table = NFert::uptake_table
)
}
\arguments{
\item{expected_yield_tons_ha, crop}{See \code{calc_crop_P_demand}.}

\item{uptake_table}{Lookup table.}
}
\value{
List with \code{K2O_requirement}, \code{unit_coef}.
}
\description{
Crop K2O demand
}
\examples{
calc_crop_K_demand(expected_yield_tons_ha = 6,
                   crop = "Grano duro (pianta intera)")
}
