% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_chemistry.R
\name{classify_carbonate_tot}
\alias{classify_carbonate_tot}
\title{Classify total carbonate (CaCO3 \%)}
\usage{
classify_carbonate_tot(
  caco3_tot,
  total_carbonate.table = nfert_data_get("total_carbonate.table")
)
}
\arguments{
\item{caco3_tot}{Numeric.}

\item{total_carbonate.table}{Lookup.}
}
\value{
List with \code{ID} and \code{class}.
}
\description{
Classify total carbonate (CaCO3 \%)
}
\examples{
classify_carbonate_tot(15)
}
