% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_chemistry.R
\name{classify_CEC}
\alias{classify_CEC}
\title{Classify cation exchange capacity (CSC, meq/100 g)
From Fert_Office v1.26 foglio Gri_K: < 10 = bassa, 10-20 = media, > 20 = alta.}
\usage{
classify_CEC(csc_meq)
}
\arguments{
\item{csc_meq}{Numeric meq/100 g.}
}
\value{
List with \code{class}.
}
\description{
Classify cation exchange capacity (CSC, meq/100 g)
From Fert_Office v1.26 foglio Gri_K: < 10 = bassa, 10-20 = media, > 20 = alta.
}
\examples{
classify_CEC(18)
}
