% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_chemistry.R
\name{classify_pH}
\alias{classify_pH}
\title{Classify soil pH}
\usage{
classify_pH(pH, ph.table = NFert::ph.table)
}
\arguments{
\item{pH}{Numeric.}

\item{ph.table}{Lookup (default \code{NFert::ph.table}).}
}
\value{
List with \code{ID_pH} and \code{class}.
}
\description{
DPI 2026 7-class scheme from Fert_Office v1.26 foglio \code{pH}:
\itemize{
\item fortemente acido: pH < 5.4
\item acido: 5.4 - 6.0
\item leggermente acido: 6.1 - 6.7
\item neutro: 6.8 - 7.3
\item leggermente alcalino: 7.4 - 8.1
\item alcalino: 8.2 - 8.6
\item fortemente alcalino: > 8.6
}
}
\examples{
classify_pH(8.3)
}
