% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-nfert_data.R
\name{nfert_normalize_oxygen}
\alias{nfert_normalize_oxygen}
\title{Map legacy oxygen / drainage labels to \code{ca.table$availability} (Slow / Normal / Fast).}
\usage{
nfert_normalize_oxygen(x)
}
\description{
Map legacy oxygen / drainage labels to \code{ca.table$availability} (Slow / Normal / Fast).
}
\keyword{internal}
