% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_N_from_crop_residues.R
\name{nitrogen_from_previous_crop_residues}
\alias{nitrogen_from_previous_crop_residues}
\title{Nitrogen from previous crop residues}
\usage{
nitrogen_from_previous_crop_residues(
  previous_crop = "Winter cereals straw removal",
  e.table = NFert::e.table
)
}
\arguments{
\item{previous_crop}{The name of the previous crop. This must match a value
in the \code{previous_crop} column of the \code{e.table}.}

\item{e.table}{A data frame containing the nitrogen values (in kg/ha) for different
previous crops. It should have at least two columns: \code{previous_crop} (character)
and \code{N} (numeric). The default is \code{NFert::e.table}.}
}
\value{
The estimated amount of nitrogen (N) in kg/ha available from the previous crop residues.
}
\description{
Estimates the amount of nitrogen (N) available from residues of a previous crop.
}
\details{
\itemize{
\item If the value in the \code{e.table} is negative, the function returns 0 (as negative N contribution
from residues is not meaningful).
\item If there is no matching entry for \code{previous_crop} in the \code{e.table}, the function returns \code{NA}
to indicate missing data.
}
}
\examples{
E <- nitrogen_from_previous_crop_residues(previous_crop = "Winter cereals straw removal")
print(E)

}
