% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_chemistry.R
\name{ratio_K_CEC}
\alias{ratio_K_CEC}
\title{K/CSC ratio assessment (Fert_Office foglio Gri_K)
<0.1\% basso; 0.1-2.1 ottimale; 2.1-5.1 alto; ... (percentage of CSC)}
\usage{
ratio_K_CEC(k_meq_per_100g, csc_meq)
}
\arguments{
\item{k_meq_per_100g}{K in meq/100g.}

\item{csc_meq}{CSC in meq/100g.}
}
\value{
List with \code{pct} and \code{class}.
}
\description{
K/CSC ratio assessment (Fert_Office foglio Gri_K)
<0.1\% basso; 0.1-2.1 ottimale; 2.1-5.1 alto; ... (percentage of CSC)
}
\examples{
ratio_K_CEC(k_meq_per_100g = 0.38, csc_meq = 18)
}
