% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_chemistry.R
\name{ratio_Mg_K}
\alias{ratio_Mg_K}
\title{Mg/K ratio assessment (Fert_Office foglio Gri_K)
0-0.1 basso; 0.1-2.1 equilibrato; 2.1-5.1 alto; >5.1 molto alto.}
\usage{
ratio_Mg_K(mg_ppm, k_ppm)
}
\arguments{
\item{mg_ppm, k_ppm}{Magnesium and potassium ppm.}
}
\value{
List with \code{ratio} and \code{class}.
}
\description{
Mg/K ratio assessment (Fert_Office foglio Gri_K)
0-0.1 basso; 0.1-2.1 equilibrato; 2.1-5.1 alto; >5.1 molto alto.
}
\examples{
ratio_Mg_K(mg_ppm = 121.5, k_ppm = 150)
}
