% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crop_N_demand.R
\name{calc_crop_N_demand}
\alias{calc_crop_N_demand}
\title{Crop requirement estimation}
\usage{
calc_crop_N_demand(
  expected_yield_tons_ha = 10,
  uptake_table = NFert::uptake_table,
  crop = "Mais trinciato (classe 700)"
)
}
\arguments{
\item{expected_yield_tons_ha}{Expected crop yield in tons per hectare (t/ha).}

\item{uptake_table}{A data frame containing unit crop absorptions for different crops.
It should have at least two columns: "crop" (character) and "N" (numeric).}

\item{crop}{The name of the crop for which to calculate the requirement. Must be a valid
value in the \code{crop} column of the \code{uptake_table}.}
}
\value{
A list containing:
\itemize{
\item N_requirement: The calculated nitrogen requirement in kilograms per hectare (kg/ha).
\item units: The units of the output value ("kg/ha").
}
}
\description{
Estimates the crop's nitrogen requirement based on unit crop absorption and expected yield.
}
\details{
The calculation is done as follows:
A = unit crop absorption (kg N/t yield) * expected production (t/ha)
}
\note{
The \code{uptake_table} should provide the unit crop absorption for the specified
\code{crop}. The default table (\code{NFert::uptake_table}) may not be suitable for all regions
or specific crop varieties.
}
