% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_N_immobilization_loss.R
\name{calc_N_immobilization_loss}
\alias{calc_N_immobilization_loss}
\title{Immobiization and dispersion}
\usage{
calc_N_immobilization_loss(B = 0, oxygen_availability = "Normal", id_rag = 1)
}
\arguments{
\item{B}{Soil fertility}

\item{id_rag}{}
}
\description{
Immobiization and dispersion
}
