% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_chemistry.R
\name{classify_carbonate_att}
\alias{classify_carbonate_att}
\title{Classify active carbonate}
\usage{
classify_carbonate_att(
  caco3_att,
  active_carbonate.table = NFert::active_carbonate.table
)
}
\arguments{
\item{caco3_att}{Numeric \%.}

\item{active_carbonate.table}{Lookup.}
}
\value{
List with \code{ID} and \code{class}.
}
\description{
Classify active carbonate
}
\examples{
classify_carbonate_att(6.8)
}
