% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_N_leaching_loss.R
\name{leaching_loss}
\alias{leaching_loss}
\title{Leaching loss nitrogen}
\usage{
leaching_loss(
  winter_rain = 160,
  start_spring_rain = 40,
  oxygen_availability = "Normal",
  id_rag = 3,
  b1 = 29.16
)
}
\arguments{
\item{winter_rain}{Rain in the winter time dal 1/10 al 31/01}

\item{start_spring_rain}{Rain in febraury}

\item{method}{}
}
\description{
Precipitation-based method (c1)
Where rainfall is concentrated in the autumn-winter period, in general,that portion of nitrogen that enters the budget as "ready N" should be considered leachable. Whereas in situations with significant rainfall surplus even during the spring-summer period and with soils with low water retention, a fraction of the nitrogen from fertilization and that from S.O. mineralization should be considered leachable in addition to ready nitrogen. Leaching losses in the autumn-winter period are estimated by taking as a reference
the amount of precipitation in the interval from October 1 to January 31 as follows below:
\itemize{
\item with rainfall <150 mm: no losses:
\item with rainfall between 150 and 250 mm: progressively increasing loss of ready nitrogen;
\item with rainfall >250 mm: all ready nitrogen is lost.
To calculate the \% of ready N that is considered to be leached as a function of rainfall, the following expression is used:
x = (y - 150)
where: x>0 = percentage of ready nitrogen lost; y = rainfall in mm during October to January.
}
}
\details{
Ease of drainage method (c2)
The calculation of nitrogen losses in soil by leaching based on drainage and texture can be estimated by adopting the following scheme.
}
