% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_N_balance.R
\name{N_balance}
\alias{N_balance}
\title{Nitrogen balance}
\usage{
N_balance(
  expected_yield_tons_ha = 15,
  crop = "Silage maize FAO 500",
  ccp = "Spring-summer crop 100–130 days",
  sand = 50,
  clay = 35,
  Ntot = 1.2,
  SOM = 1.2,
  CN = 9.5,
  oxygen_availability = "Normal",
  winter_rain = 160,
  start_spring_rain = 40,
  prev_crop = "Winter cereals straw removal",
  source = "Cattle slurry",
  fertorg_frequency = "every year",
  location = "Plain adjacent to urbanized areas",
  forg_quantity = 100
)
}
\arguments{
\item{forg_quantity}{}
}
\description{
Nitrogen balance
}
