% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_soil_N_available.R
\name{soil_fertility}
\alias{soil_fertility}
\title{Calculate soil fertility parameters.}
\usage{
soil_fertility(
  Ntot = 1.2,
  SOM = 2,
  soil.group = "Sandy textures",
  CN = 8,
  ccp = "Autumn-winter crop <150 days",
  coefN_readily = NFert::coefN_readily,
  coefN_mineralised = NFert::coefN_mineralised,
  coef_time = NFert::coef_time
)
}
\arguments{
\item{Ntot}{Total nitrogen content in the soil (kg/ha).}

\item{SOM}{Soil organic matter content (\%).}

\item{soil.group}{Soil group (e.g., "Sandy textures").}

\item{CN}{Carbon-to-nitrogen ratio.}

\item{ccp}{Crop category and planting period (e.g., "Autumn-winter crop <150 days").}
}
\value{
A named vector containing:
\itemize{
\item b1: N_readily_available: Readily available nitrogen (kg/ha).
\item b2: N_mineralized: Mineralized nitrogen (kg/ha).
\item units: The units of the output values ("kg/ha").
}
}
\description{
This function estimates the amount of readily available and mineralized nitrogen in the soil.
}
